Split out the actual node for robot_state_publisher to a separate file.
This way, joint_state_publisher can actually be used successfully
as a library.
